groq_rag/django-api/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.1
boto3==1.35.2
botocore==1.35.2
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.7
certifi==2024.2.2
chainlit==1.0.506
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cohere==5.8.1
coloredlogs==15.0.1
dataclasses-json==0.5.14
dataclasses-json-speakeasy==0.5.11
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
docstring_parser==0.16
embedchain==0.1.121
emoji==2.11.1
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.3
fastapi-socketio==0.0.10
fastavro==1.9.5
fastembed==0.2.6
filelock==3.14.0
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.29.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
gptcache==0.1.44
greenlet==3.0.3
groq==0.5.0
grpc-google-iam-v1==0.13.1
grpcio==1.63.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
instructor==1.3.3
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json-rpc==1.15.0
json_repair==0.25.3
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonref==1.1.0
jsonrpcclient==4.0.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.2.14
langchain-cohere==0.1.9
langchain-community==0.2.12
langchain-core==0.2.34
langchain-experimental==0.0.64
langchain-groq==0.1.3
langchain-openai==0.1.22
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.2.12
langgraph-checkpoint==1.0.5
langserve==0.1.1
langsmith==0.1.99
Lazify==0.4.0
literalai==0.0.509
llama-index==0.10.34
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.34
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.20
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
loguru==0.7.2
lxml==5.2.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.0.20
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.0
onnxruntime==1.17.3
openai==1.42.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
portalocker==2.10.1
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
PyJWT==2.8.0
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.11.0
rapidfuzz==3.9.0
redis==5.0.8
referencing==0.35.1
regex==2024.4.28
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.2
schema==0.7.7
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.1
sse-starlette==2.1.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.34.0
striprtf==0.0.26
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
unstructured==0.13.6
unstructured-client==0.18.0
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
